ui: move vector to util folder
Release note: None
